msm: camera: ispif: Use proper type while comparing negative values.
Not using the proper type to compare against negative values
will not work in case of 64bit build because long and int types
have different ranges for LP64 builds. So fix it by using a proper
cast.

+++ drivers/media/platform/msm/camera_v2/ispif/msm_ispif.c
@@ -194,7 +194,7 @@ int msm_ispif_get_ahb_clk_info(struct ispif_device *ispif_dev,
 	}
 	for (i = 0; i < count; i++) {
 		ispif_8974_ahb_clk_info[i].clk_rate =
-			 (rates[i] == 0) ? -1 : rates[i];
+			 (rates[i] == 0) ? (long)-1 : rates[i];
 		CDBG("clk_rate[%d] = %ld\n", i,
 			 ispif_8974_ahb_clk_info[i].clk_rate);
 	}
